// server/src/interfaces/IMetricOringAndGlandDetails.ts
export interface IMetricOringAndGlandDetails {
  _id: string;
  dashSize: number;
  oringCrossSectionSize: number;
  oringCrossSectionTolerance: number;
  oringDiameterSize: number;
  oringDiameterTolerance: number;
  internalVacuumGlandIdSize: number;
  internalVacuumGlandIdTolerance: number;
  internalPressureGlandIdSize: number;
  internalPressureGlandIdTolerance: number;
  dovetailGlandClDiameterSize: number;
  dovetailGlandClDiameterTolerance: number;
}